refactor(workers): rename extend to extend-rest, drop dead diagnostic block
